backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pillow
pytesseract
PyMuPDF
scikit-learn
pandas
joblib